backend/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Machine Learning
# Pin torch to match the GPU base image (pytorch/pytorch:2.2.0-cuda11.8)
torch==2.2.0
transformers==4.37.2
safetensors>=0.4.0
tokenizers>=0.15.0
sentencepiece>=0.1.99
protobuf>=3.20.0
pillow>=10.0.0
einops>=0.7.0
timm>=0.9.12
# flash-attn>=2.5.0  # Optional - Florence-2 works without it (uses eager attention)

# Configuration and environment
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# File handling
python-multipart>=0.0.6

# Development and logging
python-json-logger>=2.0.0

# Optional: For better performance
accelerate>=0.27.0
huggingface-hub>=0.21.0

# Health checks
requests>=2.31.0